Better [#CLICKHOUSE-2]

+++ dbms/src/Storages/MergeTree/MergeTreeDataPart.h
@@ -21,16 +21,16 @@ struct MergeTreeDataPartChecksum
 {
     using uint128 = CityHash_v1_0_2::uint128;
 
-    size_t file_size {};
+    UInt64 file_size {};
     uint128 file_hash {};
 
     bool is_compressed = false;
-    size_t uncompressed_size {};
+    UInt64 uncompressed_size {};
     uint128 uncompressed_hash {};
 
     MergeTreeDataPartChecksum() {}
-    MergeTreeDataPartChecksum(size_t file_size_, uint128 file_hash_) : file_size(file_size_), file_hash(file_hash_) {}
-    MergeTreeDataPartChecksum(size_t file_size_, uint128 file_hash_, size_t uncompressed_size_, uint128 uncompressed_hash_)
+    MergeTreeDataPartChecksum(UInt64 file_size_, uint128 file_hash_) : file_size(file_size_), file_hash(file_hash_) {}
+    MergeTreeDataPartChecksum(UInt64 file_size_, uint128 file_hash_, UInt64 uncompressed_size_, uint128 uncompressed_hash_)
         : file_size(file_size_), file_hash(file_hash_), is_compressed(true),
         uncompressed_size(uncompressed_size_), uncompressed_hash(uncompressed_hash_) {}
 
@@ -50,7 +50,7 @@ struct MergeTreeDataPartChecksums
     using FileChecksums = std::map<String, Checksum>;
     FileChecksums files;
 
-    void addFile(const String & file_name, size_t file_size, Checksum::uint128 file_hash);
+    void addFile(const String & file_name, UInt64 file_size, Checksum::uint128 file_hash);
 
     void add(MergeTreeDataPartChecksums && rhs_checksums);
 
@@ -104,10 +104,10 @@ struct MergeTreeDataPart
     const Checksum * tryGetMrkChecksum(const String & name) const;
 
     /// Returns the size of .bin file for column `name` if found, zero otherwise
-    size_t getColumnCompressedSize(const String & name) const;
-    size_t getColumnUncompressedSize(const String & name) const;
+    UInt64 getColumnCompressedSize(const String & name) const;
+    UInt64 getColumnUncompressedSize(const String & name) const;
     /// Returns the size of .mrk file for column `name` if found, zero otherwise
-    size_t getColumnMrkSize(const String & name) const;
+    UInt64 getColumnMrkSize(const String & name) const;
 
     /// Returns the name of a column with minimum compressed size (as returned by getColumnSize()).
     /// If no checksums are present returns the name of the first physically existing column.
@@ -136,7 +136,7 @@ struct MergeTreeDataPart
 
     size_t rows_count = 0;
     size_t marks_count = 0;
-    std::atomic<size_t> size_in_bytes {0};  /// size in bytes, 0 - if not counted;
+    std::atomic<UInt64> size_in_bytes {0};  /// size in bytes, 0 - if not counted;
                                             ///  is used from several threads without locks (it is changed with ALTER).
     time_t modification_time = 0;
     mutable time_t remove_time = std::numeric_limits<time_t>::max(); /// When the part is removed from the working set.
@@ -255,7 +255,7 @@ struct MergeTreeDataPart
     /// Columns description.
     NamesAndTypesList columns;
 
-    using ColumnToSize = std::map<std::string, size_t>;
+    using ColumnToSize = std::map<std::string, UInt64>;
 
     /** It is blocked for writing when changing columns, checksums or any part files.
         * Locked to read when reading columns, checksums or any part files.
@@ -275,7 +275,7 @@ struct MergeTreeDataPart
     ~MergeTreeDataPart();
 
     /// Calculate the total size of the entire directory with all the files
-    static size_t calculateTotalSize(const String & from);
+    static UInt64 calculateTotalSize(const String & from);
 
     void remove() const;
 
@@ -297,10 +297,10 @@ struct MergeTreeDataPart
     bool hasColumnFiles(const String & column) const;
 
     /// For data in RAM ('index')
-    size_t getIndexSizeInBytes() const;
-    size_t getIndexSizeInAllocatedBytes() const;
+    UInt64 getIndexSizeInBytes() const;
+    UInt64 getIndexSizeInAllocatedBytes() const;
     /// Total size of *.mrk files
-    size_t getTotalMrkSizeInBytes() const;
+    UInt64 getTotalMrkSizeInBytes() const;
 
 private:
     /// Reads columns names and types from columns.txt
